[feature] SSEClientTransport support w/ mock server tests
Finished the SSEClientTransport class for connecting to SSE-based MCP
servers. This handles creation of the downchannel and serializing of the
POST messages to the negotiated server endpointURL. The endpointURL is
decided with the `endpoint` event during initial client downchannel connection.

We bubble errors on the `transport.messages` iterator when data fails to
be parsed, or when the stream fails unexpectedly (ex: network error).
We dont throw when the connection ends due to client or server
termination; instead we fail gracefully and the continuations /
iterators are just closed.

I also added a js-based SSE MCP server for testing the class. This is
spawned by the unit test. Your computer must have node accessible from
your PATH in order to run the tests. This dummy server simulates data
being sent, as well as some custom error scenarios for testing the
client.

I added tests for the basic logic of the transport class. Ill work on
adding this to Stdio transport classes next.